Introduce mock controller that serves both the destination and identity services
